# backend/docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  models.Product:
    properties:
      cost:
        type: number
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      id:
        type: integer
      inventory:
        type: integer
      list_price:
        type: number
      name:
        type: string
      product_number:
        type: string
      supplier:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: EZStore B2B 報價與訂單管理系統 API。提供客戶管理、產品管理、報價單管理（含報價因子計算）及訂單管理功能。
  title: EZStore API
  version: "1.0"
paths:
...
  /api/v1/products:
    get:
      consumes:
      - application/json
      description: 取得產品列表，支援分頁與搜尋（搜尋產品名稱或產品編號）
      parameters:
      - default: 1
        description: 頁碼（預設 1）
        in: query
        name: page
        type: integer
      - default: 20
        description: 每頁筆數（預設 20）
        in: query
        name: page_size
        type: integer
      - description: 搜尋關鍵字
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 列出產品
      tags:
      - products
    post:
      consumes:
      - application/json
      description: 建立新產品，需提供產品編號、名稱、成本、牌價等資料
      parameters:
      - description: 產品資料
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 建立產品
      tags:
      - products
  /api/v1/products/{id}:
    delete:
      consumes:
      - application/json
      description: 根據 ID 軟刪除產品
      parameters:
      - description: 產品 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 刪除產品
      tags:
      - products
    get:
      consumes:
      - application/json
      description: 根據 ID 取得產品詳細資料，包含成本、牌價、庫存等資訊
      parameters:
      - description: 產品 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: 取得單一產品
      tags:
      - products
    put:
      consumes:
      - application/json
      description: 根據 ID 修改產品資料
      parameters:
      - description: 產品 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 產品資料
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/models.Product'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: 修改產品
      tags:
      - products
...
swagger: "2.0"